moved sidebar to SideBar.jsx
